Add probably-unsound "smugglers mutex"
For passing non-'static data from a Bevy system
to a Wasmtime host-defined function.

It seems to "work" in my limited testing. I'll see if I can
convince myself that it's sound or do what I need to make it so,
and then figure out how the heck to make it generic.